Feth data on page load added
