+++ src/index.html
@@ -21,11 +21,6 @@
                 <button id="refresh-score" class="btn btn-lg btn-outline-success">Refresh</button>
             </div>
             <ul id="list-scores" class="list-group my-3">
-                <li class="list-group-item list-group-item-action">Name: 100</li>
-                <li class="list-group-item list-group-item-action">Name: 20</li>
-                <li class="list-group-item list-group-item-action">Name: 50</li>
-                <li class="list-group-item list-group-item-action">Name: 78</li>
-                <li class="list-group-item list-group-item-action">Name: 100</li>
               </ul>
         </section>
         <section id="add-scores" class="col-md-6 my-3">
