+++ src/index.js
@@ -3,7 +3,7 @@ import SaveToLocalStorage from './modules/Game/SaveIdGame.js';
 import ReadLocalStorage from './modules/Game/ReadIdGame.js';
 import getGameId from './api/Game.js';
 import CreateScore from './modules/Scores/CreateScore.js';
-import ReadScores from './modules/Scores/ReadScores.js';
+import DisplayScores from './modules/Scores/DisplayScores.js';
 
 /* SaveToLocalStorage('asdsasadasd'); */
 let gameId = null;
@@ -21,29 +21,15 @@ gameId = currentIdGame?.id ? currentIdGame.id : getId();
 // Submit form
 const form = document.getElementById('add-score-form');
 
-form.addEventListener('submit', async (e) => {
+form.addEventListener('submit', (e) => {
   e.preventDefault();
-  await CreateScore(gameId);
+  CreateScore(gameId);
 });
 
 // Get recorded data
 const refreshScore = document.getElementById('refresh-score');
-refreshScore.addEventListener('click', async () => {
-  const listScores = document.getElementById('list-scores');
-
-  // Remove previous data
-  const prevLi = listScores.querySelectorAll('li');
-  prevLi.forEach((li) => {
-    li.remove();
-  });
-
-  const response = await ReadScores(gameId);
-  response?.result.map((record) => {
-    const { user, score } = record;
-    const liScore = document.createElement('li');
-    liScore.classList.add('list-group-item', 'list-group-item-action');
-    liScore.textContent = user + score;
-
-    return listScores.appendChild(liScore);
-  });
+refreshScore.addEventListener('click', () => {
+  DisplayScores(gameId);
 });
+
+DisplayScores(gameId);
